Update text import
